+++ packages/core/streams/ai-stream.ts
@@ -5,110 +5,173 @@ import {
   type ReconnectInterval
 } from 'eventsource-parser'
 
+/**
+ * Helper callback methods for AIStream stream lifecycle events
+ * @interface
+ */
 export interface AIStreamCallbacks {
   onStart?: () => Promise<void>
   onCompletion?: (completion: string) => Promise<void>
   onToken?: (token: string) => Promise<void>
 }
 
+/**
+ * Custom parser for AIStream data.
+ * @interface
+ */
 export interface AIStreamParser {
   (data: string): string | void
 }
 
-export function createEventStreamTransformer(customParser: AIStreamParser) {
-  const decoder = new TextDecoder()
-  let parser: EventSourceParser
+/**
+ * Creates a TransformStream that parses events from an EventSource stream using a custom parser.
+ * @param {AIStreamParser} customParser - Function to handle event data.
+ * @returns {TransformStream<Uint8Array, string>} TransformStream parsing events.
+ */
+export function createEventStreamTransformer(
+  customParser: AIStreamParser
+): TransformStream<Uint8Array, string> {
+  const textDecoder = new TextDecoder()
+  let eventSourceParser: EventSourceParser
 
-  return new TransformStream<Uint8Array, string>({
+  return new TransformStream({
     async start(controller): Promise<void> {
-      function onParse(event: ParsedEvent | ReconnectInterval) {
-        if (event.type === 'event') {
-          const data = event.data
-          if (data === '[DONE]') {
+      eventSourceParser = createParser(
+        (event: ParsedEvent | ReconnectInterval) => {
+          if (
+            'data' in event &&
+            event.type === 'event' &&
+            event.data === '[DONE]'
+          ) {
             controller.terminate()
             return
           }
 
-          const message = customParser(data)
-          if (message) controller.enqueue(message)
+          if ('data' in event) {
+            const parsedMessage = customParser(event.data)
+            if (parsedMessage) controller.enqueue(parsedMessage)
+          }
         }
-      }
-
-      parser = createParser(onParse)
+      )
     },
 
     transform(chunk) {
-      parser.feed(decoder.decode(chunk))
+      eventSourceParser.feed(textDecoder.decode(chunk))
     }
   })
 }
 
 /**
- * This stream forks input stream, allowing us to use the result as a
- * bytestream of the messages and pass the messages to our callback interface.
+ * Creates a transform stream that encodes input messages and invokes optional callback functions.
+ * The transform stream uses the provided callbacks to execute custom logic at different stages of the stream's lifecycle.
+ * - `onStart`: Called once when the stream is initialized.
+ * - `onToken`: Called for each tokenized message.
+ * - `onCompletion`: Called once when the stream is flushed, with the aggregated messages.
+ *
+ * This function is useful when you want to process a stream of messages and perform specific actions during the stream's lifecycle.
+ *
+ * @param {AIStreamCallbacks} [callbacks] - An object containing the callback functions.
+ * @return {TransformStream<string, Uint8Array>} A transform stream that encodes input messages as Uint8Array and allows the execution of custom logic through callbacks.
+ *
+ * @example
+ * const callbacks = {
+ *   onStart: async () => console.log('Stream started'),
+ *   onToken: async (token) => console.log(`Token: ${token}`),
+ *   onCompletion: async (completion) => console.log(`Completion: ${completion}`)
+ * };
+ * const transformer = createCallbacksTransformer(callbacks);
  */
 export function createCallbacksTransformer(
   callbacks: AIStreamCallbacks | undefined
-) {
-  const encoder = new TextEncoder()
-  let fullResponse = ''
-
+): TransformStream<string, Uint8Array> {
+  const textEncoder = new TextEncoder()
+  let aggregatedResponse = ''
   const { onStart, onToken, onCompletion } = callbacks || {}
 
-  return new TransformStream<string, Uint8Array>({
+  return new TransformStream({
     async start(): Promise<void> {
       if (onStart) await onStart()
     },
 
     async transform(message, controller): Promise<void> {
-      controller.enqueue(encoder.encode(message))
+      controller.enqueue(textEncoder.encode(message))
 
       if (onToken) await onToken(message)
-      if (onCompletion) fullResponse += message
+      if (onCompletion) aggregatedResponse += message
     },
 
     async flush(): Promise<void> {
-      await onCompletion?.(fullResponse)
+      if (onCompletion) await onCompletion(aggregatedResponse)
     }
   })
 }
 
-// If we're still at the start of the stream, we want to trim the leading
-// `\n\n`. But, after we've seen some text, we no longer want to trim out
-// whitespace.
-export function trimStartOfStreamHelper() {
-  let start = true
-  return (text: string) => {
-    if (start) text = text.trimStart()
-    if (text) start = false
+/**
+ * Returns a stateful function that, when invoked, trims leading whitespace
+ * from the input text. The trimming only occurs on the first invocation, ensuring that
+ * subsequent calls do not alter the input text. This is particularly useful in scenarios
+ * where a text stream is being processed and only the initial whitespace should be removed.
+ *
+ * @return {function(string): string} A function that takes a string as input and returns a string
+ * with leading whitespace removed if it is the first invocation; otherwise, it returns the input unchanged.
+ *
+ * @example
+ * const trimStart = trimStartOfStreamHelper();
+ * const output1 = trimStart("   text"); // "text"
+ * const output2 = trimStart("   text"); // "   text"
+ *
+ */
+export function trimStartOfStreamHelper(): (text: string) => string {
+  let isStreamStart = true
+
+  return (text: string): string => {
+    if (isStreamStart) {
+      text = text.trimStart()
+      if (text) isStreamStart = false
+    }
     return text
   }
 }
 
+/**
+ * Returns a ReadableStream created from the response, parsed and handled with custom logic.
+ * The stream goes through two transformation stages, first parsing the events and then
+ * invoking the provided callbacks.
+ * @param {Response} response - The response.
+ * @param {AIStreamParser} customParser - The custom parser function.
+ * @param {AIStreamCallbacks} callbacks - The callbacks.
+ * @return {ReadableStream} The AIStream.
+ * @throws Will throw an error if the response is not OK.
+ */
 export function AIStream(
-  res: Response,
+  response: Response,
   customParser: AIStreamParser,
   callbacks?: AIStreamCallbacks
 ): ReadableStream {
-  // If the response is not OK, we want to throw an error to indicate that
-  // the AI service is not available.
-  // When catching this error, we can check the status code and return a handled
-  // error response to the client.
-  if (!res.ok) {
+  if (!response.ok) {
     throw new Error(
-      `Failed to convert the response to stream. Received status code: ${res.status}.`
+      `Failed to convert the response to stream. Received status code: ${response.status}.`
     )
   }
 
-  const stream =
-    res.body ||
-    new ReadableStream({
-      start(controller) {
-        controller.close()
-      }
-    })
+  const responseBodyStream = response.body || createEmptyReadableStream()
 
-  return stream
+  return responseBodyStream
     .pipeThrough(createEventStreamTransformer(customParser))
     .pipeThrough(createCallbacksTransformer(callbacks))
+}
+
+/**
+ * Creates an empty ReadableStream that immediately closes upon creation.
+ * This function is used as a fallback for creating a ReadableStream when the response body is null or undefined,
+ * ensuring that the subsequent pipeline processing doesn't fail due to a lack of a stream.
+ *
+ * @returns {ReadableStream} An empty and closed ReadableStream instance.
+ */
+function createEmptyReadableStream(): ReadableStream {
+  return new ReadableStream({
+    start(controller) {
+      controller.close()
+    }
+  })
 }